refactor: eng review cleanup — pagination, LiveTutor split, narrower exceptions
P1-3: Add query limits to teacher_analytics (5000 safety cap) and
pagination (limit/offset) to learning_errors endpoints
P2-6: Split LiveTutor.tsx (1101→720 lines) into live-tutor/ directory
with ParagraphCard, FeedbackPanel, and helpers sub-modules
P2-7: Narrow 3 bare except Exception catches in ai_base.py and
socratic_agent.py to specific types (JSON/SQLAlchemy/ValueError)

via [Happy](https://happy.engineering)
