+++ TODOS.md
@@ -20,12 +20,9 @@ Generated by `/plan-eng-review` full site review on staging branch.
 - **Depends on**: Vitest setup (check if vitest.config exists)
 - **Effort**: CC ~1 hour
 
-### 3. Add pagination to teacher analytics + learning errors
-- **What**: Add LIMIT/offset to `teacher_analytics.py` (7 unbounded .all()) and `learning_errors.py` (7 unbounded .all())
-- **Why**: Queries are classroom-scoped (~30-50 students x 57 lessons = ~1500 rows max). Low urgency for current scale, but will matter with growth
-- **How to apply**: Copy pagination pattern from `learning_sessions.py` (already has limit/offset Query params)
-- **Note**: SQL aggregation approach (extract epoch) breaks SQLite tests. Keep Python-side processing for now
-- **Effort**: CC ~20 min
+### ~~3. Add pagination to teacher analytics + learning errors~~ ✅ DONE
+- teacher_analytics: added `.limit(5000)` safety cap on unbounded session queries
+- learning_errors: added `limit`/`offset` Query params to `get_error_patterns` and configurable `limit` to `get_recommended_vocab`
 
 ## P2 — Medium Impact
 
@@ -35,17 +32,14 @@ Generated by `/plan-eng-review` full site review on staging branch.
 ### ~~5. Remove dead TenantMiddleware~~ ✅ DONE
 - Deleted `middleware/tenant.py` + `tests/test_tenant_middleware.py` + removed from main.py
 
-### 6. Split LiveTutor.tsx into sub-components
-- **What**: Extract ParagraphView, TutorControlPanel, FeedbackDisplay from LiveTutor.tsx (1,101 lines)
-- **Why**: Over 1000-line critical threshold. Hard to maintain, hard to test
-- **How to apply**: Co-located sub-components in `reading-steps/live-tutor/` directory
-- **Effort**: CC ~20 min
+### ~~6. Split LiveTutor.tsx into sub-components~~ ✅ DONE
+- Extracted to `reading-steps/live-tutor/`: helpers.ts (134), FeedbackPanel.tsx (129), ParagraphCard.tsx (341), LiveTutor.tsx (720), index.tsx (2)
+- Original LiveTutor.tsx is now a 6-line re-export shim for backward compatibility
 
-### 7. Narrow bare `except Exception` catches in AI/learning paths
-- **What**: 55 bare `except Exception` catches across backend, 16 in AI/learning paths
-- **Why**: On closer review, these are well-designed fail-safes — they all log the error and fall back gracefully (e.g., `understood=False`, canned fallback question). Circuit breaker handles repeated failures. Low priority refinement
-- **How to apply**: Replace with specific catches where it adds clarity, but don't break the fail-safe pattern
-- **Effort**: CC ~15 min
+### ~~7. Narrow bare `except Exception` catches in AI/learning paths~~ ✅ DONE
+- ai_base.py: `_repair_json` narrowed to `(json.JSONDecodeError, ValueError)`
+- socratic_agent.py: DB session restore → `(SQLAlchemyError, AttributeError, KeyError)`, DB delete → `SQLAlchemyError`
+- 10 catches left as `except Exception` — intentional fail-safes around external AI calls
 
 ## P3 — Low Priority (noted, no action needed now)
