+++ backend/app/routes/learning/learning_errors.py
@@ -122,6 +122,8 @@ def get_student_story_slugs(
 def get_error_patterns(
     student_id: int,
     story_slug: str | None = Query(None, description="Filter errors by story slug"),
+    limit: int = Query(50, ge=1, le=200, description="Max patterns to return"),
+    offset: int = Query(0, ge=0, description="Number of patterns to skip"),
     current_user: User = Depends(get_current_user),
     db: Session = Depends(get_db),
 ):
@@ -170,10 +172,10 @@ def get_error_patterns(
     for row in mastered_rows:
         mastered_chars.add(row.character)
 
-    patterns = []
+    all_patterns = []
     for row in error_groups:
         is_corrected = row.character in mastered_chars
-        patterns.append(ErrorPatternItem(
+        all_patterns.append(ErrorPatternItem(
             character=row.character,
             total_error_count=row.total_error_count,
             sessions_with_error=row.sessions_with_error,
@@ -182,12 +184,15 @@ def get_error_patterns(
             is_corrected=is_corrected,
         ))
 
-    return ErrorPatternsResponse(patterns=patterns, total=len(patterns))
+    total = len(all_patterns)
+    patterns = all_patterns[offset:offset + limit]
+    return ErrorPatternsResponse(patterns=patterns, total=total)
 
 
 @router.get("/learning/students/{student_id}/recommended-vocab", response_model=RecommendedVocabResponse)
 def get_recommended_vocab(
     student_id: int,
+    limit: int = Query(10, ge=1, le=50, description="Max vocab items to return"),
     current_user: User = Depends(get_current_user),
     db: Session = Depends(get_db),
 ):
@@ -231,7 +236,7 @@ def get_recommended_vocab(
     for row in error_groups:
         if row.character in mastered_chars:
             continue
-        if len(items) >= 10:
+        if len(items) >= limit:
             break
         items.append(RecommendedVocabItem(
             character=row.character,
