+++ frontend/src/components/reading-steps/live-tutor/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './LiveTutor';
+export type { LineResult } from './helpers';
